ref(js): Colocate api error handling code to api.tsx
